# BizzyFrontend/contracts/backend.openapi.json
{
  "components": {
    "schemas": {
...
      "Evidence": {
        "properties": {
          "dimensions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Dimensions",
            "type": "object"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "metric",
          "value"
        ],
        "title": "Evidence",
        "type": "object"
      },
      "FinanceReport": {
        "properties": {
          "as_of_date": {
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "currency": {
            "const": "SGD",
            "default": "SGD",
            "title": "Currency",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "report_type": {
            "enum": [
              "profitability",
              "operating_expenses",
              "cash_flow",
              "receivables"
            ],
            "title": "Report Type",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "report_type",
          "start_date",
          "as_of_date",
          "summary",
          "evidence"
        ],
        "title": "FinanceReport",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BizzyBee Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/finance/report": {
      "get": {
        "operationId": "finance_report_api_v1_finance_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "report_type",
            "required": true,
            "schema": {
              "enum": [
                "profitability",
                "operating_expenses",
                "cash_flow",
                "receivables"
              ],
              "title": "Report Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "as_of_date",
            "required": false,
            "schema": {
              "default": "2026-09-23",
              "format": "date",
              "title": "As Of Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finance Report",
        "tags": [
          "bizzybee"
        ]
      }
    },
...
  }
}